backend/requirements.txt
# --- FastAPI ---
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# --- LangChain (correct & compatible) ---
langchain==1.1.3
langchain-core==1.1.3
langchain-community==0.4.1
langchain-aws==1.1.0

# --- Vector DB ---
chromadb==0.4.24
langchain-chroma==0.2.0

# --- Documents ---
pypdf==4.2.0

# --- Utilities ---
python-dotenv==1.0.0
requests

# --- Data ---
pydantic==2.12.5
sqlalchemy==2.0.25
numpy<2.0